+++ crates/spirv-std/src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
+#![feature(adt_const_params)]
 #![cfg_attr(
     target_arch = "spirv",
     allow(internal_features),