+++ tests/compiletests/ui/arch/atomic_i_increment.rs
@@ -10,8 +10,8 @@ pub fn main(#[spirv(descriptor_set = 0, binding = 0, storage_buffer)] buffer: &m
     let old = unsafe {
         spirv_std::arch::atomic_i_increment::<
             _,
-            { spirv_std::memory::Scope::Workgroup as u32 },
-            { spirv_std::memory::Semantics::NONE.bits() as u32 },
+            { spirv_std::memory::Scope::Workgroup },
+            { spirv_std::memory::Semantics::NONE },
         >(reference)
     };
     assert!(old == 0);